+++ CHANGES.md
@@ -6,6 +6,10 @@ Version 0.41.4
 
 To be released.
 
+ -  (Libplanet.Explorer) Fixed a bug of `TransactionQuery<T>` that
+    `bindSignature()` had errored if `unsignedTransaction` has a system
+    action.  [[#2358]]
+
 
 Version 0.41.3
 --------------
@@ -203,13 +207,25 @@ Released on August 26, 2022.
 [#2241]: https://github.com/planetarium/libplanet/pull/2241
 
 
+Version 0.40.3
+--------------
+
+Released on September 29, 2022.
+
+ -  (Libplanet.Explorer) Fixed a bug of `TransactionQuery<T>` that
+    `bindSignature()` had errored if `unsignedTransaction` has a system
+    action.  [[#2358]]
+
+[#2358]: https://github.com/planetarium/libplanet/pull/2358
+
+
 Version 0.40.2
 --------------
 
 Released on September 23, 2022.
 
--  Fixed a bug where `Transaction<T>.ToBencodex(bool)` method had incorrectly
-   serialized `SystemAction`.  [[#2339]]
+ -  Fixed a bug where `Transaction<T>.ToBencodex(bool)` method had incorrectly
+    serialized `SystemAction`.  [[#2339]]
 
 [#2339]: https://github.com/planetarium/libplanet/pull/2339